backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic